Move angle between vectors to utils
